requirements.txt
Django==1.4.5
Djapian==2.3.1
Markdown==2.0.3
PIL==1.1.7
Pillow==1.7.8
Pygments==1.6
South==0.7.6
Whoosh==2.4.1
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
celery==3.0.9
celery-haystack==0.6.2
distribute==0.6.34
dj-database-url==0.2.1
django-annoying==0.7.7
django-appconf==0.6
django-authopenid==1.0.1
django-autoslug==1.6.1
django-bootstrap-form==2.0.6
django-braces==1.0.0
django-category==0.1.1
django-celery-transactions==0.1.3
django-cities==0.2
django-cities-light==2.0.7
django-classy-tags==0.4
django-common==0.1.51
django-compressor==1.3
django-countries==1.5
django-daddy-avatar==1.0.3
django-debug-toolbar==0.9.4
django-endless-pagination==2.0
django-extensions==1.0.3
django-guardian==1.1.0
django-haystack==2.0.0
django-images==0.3
django-mailer==0.1.0
django-messages==0.4.4
django-mptt==0.5.2
django-mptt-admin==0.1.3
django-notification==1.0
django-pagedown==0.0.4
django-pagination==1.0.7
django-pin==1.4.5
django-polymorphic==0.5
django-polymorphic-tree==0.8.7
django-registration==0.8
django-shop==0.2.0
django-shop-categories==0.8.beta.2
django-sitetree==0.9.4
django-staticfiles-masonry==2.1.05
django-tag-parser==1.0.0
django-tagging==0.3.1
django-tagging-autocomplete==0.3.1
django-taggit==0.10a1
django-tastypie==0.9.12
django-treeadmin==0.4
django-user-profile==1.0.3
django-userena==1.2.0
django-wmd-editor==0.9.0
django-wpadmin==0.2.0
easy-thumbnails==1.2
gunicorn==0.14.6
jsonfield==0.9.15
kombu==2.5.10
lxml==3.1.0
markdown2==2.1.0
mimeparse==0.1.3
postmarkup==1.2.0
progressbar==2.2
psycopg2==2.4.6
pyCalverter==1.6.1
python-dateutil==1.5
python-openid==2.2.5
pytils==0.2.3
pytz==2012h
requests==1.2.3
six==1.3.0
sorl-thumbnail==11.09
typogrify==1.0
versiontools==1.9.1
wsgiref==0.1.2
xapian-haystack==1.1.5beta
git+git://github.com/strange/django-country-utils.git
git+git://github.com/pije76/django-tagging-autocomplete-tag-it.git
